requirements.txt
fastapi
uvicorn
python-dotenv
elevenlabs
